requirements.txt
feedparser
google-genai
chromadb
requests
python-telegram-bot
gtts
edge-tts
streamlit
pyngrok
nest_asyncio
python-dotenv